+++ src/main/java/dev/muffin/rpgcore/rpg/utils/RPGStats.java
@@ -1,21 +1,21 @@
 package dev.muffin.rpgcore.rpg.utils;
 
 public class RPGStats {
-    private int mana;
+    private double mana;
     private double armor;
     private double magicResist;
 
-    public RPGStats(int mana, double armor, double magicResist) {
+    public RPGStats(double mana, double armor, double magicResist) {
         this.mana = mana;
         this.armor = armor;
         this.magicResist = magicResist;
     }
 
-    public int getMana() {
+    public double getMana() {
         return mana;
     }
 
-    public void setMana(int mana) {
+    public void setMana(double mana) {
         this.mana = mana;
     }
